refactor: used hooks for abstraction
