+++ src/components/Editor/Editor.jsx
@@ -8,14 +8,13 @@ import {
   BsX,
   BsArchiveFill,
 } from "react-icons/bs";
-import { useLocation, useNavigate } from "react-router-dom";
-import { updateNote } from "../../services";
+import { useLocation } from "react-router-dom";
 import { useUser } from "../../context/Context";
-import { moveToArchive, restoreFromArchive } from "../../services/archive";
+import { useHandlers } from "../../hooks/useHandlers";
+import { themes } from "../../utils/utils";
 export const Editor = () => {
   const {
     user: { archives },
-    setUser,
     pinned,
     setPin,
     handlers: { updateNotePinHandler },
@@ -23,117 +22,20 @@ export const Editor = () => {
   const location = useLocation();
   const { note: oldNote } = location.state;
   const editorRef = useRef(null);
-  const navigate = useNavigate();
-
   const [themeColor, setThemeColor] = useState("#1b1b1e");
-  const [tags, setTags] = useState([]);
-  const [priority, setPriority] = useState("");
 
-  const themes = [
-    {
-      name: "Purple",
-      hexCode: "#242038",
-    },
-    {
-      name: "Prussian Blue",
-      hexCode: "#0F172A",
-    },
-    {
-      name: "Earle Black",
-      hexCode: "#1b1b1e",
-    },
-    {
-      name: "Maroon",
-      hexCode: "#96031A",
-    },
-    {
-      name: "Coffee Brown",
-      hexCode: "#1C1917",
-    },
-  ];
-
-  const updateNoteHandler = async (textContent) => {
-    let [title, ...content] = textContent.split(`\n`);
-    // sometimes the first line is empty when user starts typing from second line. so weare moving the first line to second line
-    if (title === "" && content.length) {
-      title = content[0];
-      content = content.slice(1);
-    }
-    const contentString = content.join("\r\n");
-    const { note } = await updateNote({
-      title,
-      content: contentString,
-      themeColor,
-      tags,
-      _id: oldNote._id,
-    });
-    setUser((user) => ({
-      ...user,
-      notes: user.notes.map((node) => (node._id === oldNote._id ? note : node)),
-    }));
-  };
-
-  const moveToArchivesHandler = async (id) => {
-    const { archives, status } = await moveToArchive(id);
-    setUser((user) => ({
-      ...user,
-      notes: user.notes.filter((note) => note._id !== id),
-      archives,
-    }));
-    if (status === 200) navigate("/");
-  };
-
-  const updateNoteColorHandler = async (color) => {
-    const { note } = await updateNote({
-      themeColor: color,
-      _id: oldNote._id,
-    });
-    setUser((user) => ({
-      ...user,
-      notes: user.notes.map((node) => (node._id === oldNote._id ? note : node)),
-    }));
-    setThemeColor(color);
-  };
-
-  const updateNoteTagsHandler = async (tag, isRemove = false) => {
-    console.log(
-      tag,
-      isRemove,
-      tags.filter((prev) => prev !== tag)
-    );
-    const { note } = await updateNote({
-      tags: isRemove ? tags.filter((prev) => prev !== tag) : [...tags, tag],
-      _id: oldNote._id,
-    });
-    setUser((user) => ({
-      ...user,
-      notes: user.notes.map((node) => (node._id === oldNote._id ? note : node)),
-    }));
-    setTags(note.tags);
-  };
-
-  const updateNotePriorityHandler = async (priority) => {
-    const { note } = await updateNote({
-      priority,
-      _id: oldNote._id,
-    });
-
-    setUser((user) => ({
-      ...user,
-      notes: user.notes.map((node) => (node._id === oldNote._id ? note : node)),
-    }));
-    setPriority(priority);
-  };
-
-  const removeFromArchiveHandler = async (id) => {
-    const { archives, notes, status } = await restoreFromArchive(id);
-    setUser((user) => ({
-      ...user,
-      notes,
-      archives,
-    }));
-    if (status === 200) navigate("/");
-  };
+  const {
+    tags,
+    setTags,
+    priority,
+    setPriority,
+    updateNoteHandler,
+    updateNoteColorHandler,
+    updateNotePriorityHandler,
+    updateNoteTagsHandler,
+    removeFromArchiveHandler,
+    moveToArchivesHandler,
+  } = useHandlers();
 
   const debounce = (func, delay) => {
     let inDebounce;
@@ -163,7 +65,9 @@ export const Editor = () => {
         <select
           className="editor__priority input"
           value={priority}
-          onChange={(e) => updateNotePriorityHandler(e.target.value)}
+          onChange={(e) =>
+            updateNotePriorityHandler(e.target.value, oldNote._id)
+          }
         >
           <option value="">Priority</option>
           <option value="low">Low</option>
@@ -202,7 +106,13 @@ export const Editor = () => {
         className="editor__textarea input p-lg fs-m"
         defaultValue={oldNote.title + "\r\n" + oldNote.content}
         placeholder="Start Typing..."
-        onChange={(e) => debounced(e.target.value)}
+        onChange={(e) =>
+          debounced({
+            textContent: e.target.value,
+            themeColor,
+            _id: oldNote._id,
+          })
+        }
       ></textarea>
       <footer className="editor__footer flex p-xs flex-center spread">
         <ul className="editor__tags flex flex-center list">
@@ -212,7 +122,7 @@ export const Editor = () => {
               <BsX
                 size={15}
                 className="pos-abs rounded-m pointer"
-                onClick={() => updateNoteTagsHandler(tag, true)}
+                onClick={() => updateNoteTagsHandler(tag, true, oldNote._id)}
               />
             </li>
           ))}
@@ -237,7 +147,13 @@ export const Editor = () => {
               title={theme.name}
               key={theme.hexCode}
               style={{ backgroundColor: theme.hexCode }}
-              onClick={() => updateNoteColorHandler(theme.hexCode)}
+              onClick={() =>
+                updateNoteColorHandler(
+                  oldNote._id,
+                  theme.hexCode,
+                  setThemeColor
+                )
+              }
             ></li>
           ))}
         </ul>
