+++ src/services/index.js
@@ -1,2 +1,3 @@
 export * from "./auth";
 export * from "./note";
+export * from "./archive";
