+++ src/utils/utils.js
@@ -1,0 +1,22 @@
+export const themes = [
+  {
+    name: "Purple",
+    hexCode: "#242038",
+  },
+  {
+    name: "Prussian Blue",
+    hexCode: "#0F172A",
+  },
+  {
+    name: "Earle Black",
+    hexCode: "#1b1b1e",
+  },
+  {
+    name: "Maroon",
+    hexCode: "#96031A",
+  },
+  {
+    name: "Coffee Brown",
+    hexCode: "#1C1917",
+  },
+];
